Docker/requirements.txt
<<<<<<< HEAD
tensorflow-cpu==2.18.0
tensorflow-datasets==4.9.4
numpy==1.26.4
matplotlib==3.8.2
scikit-learn==1.4.0
=======
tensorflow-cpu==2.18.0
tensorflow-datasets==4.9.4
numpy==1.26.4
matplotlib==3.8.2
scikit-learn==1.4.0
>>>>>>> 16f0156 (Initial commit - EfficientNet Pneumonia Classification with Docker)
medmnist==2.2.3